% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_peaks.R
\name{consensus_peaks}
\alias{consensus_peaks}
\title{Consensus peaks}
\usage{
consensus_peaks(peaks)
}
\arguments{
\item{peaks}{list of GRanges objects representing the peaks for each sample}
}
\value{
GRanges, the consensus peak set
}
\description{
generate a set of consensus peaks from per-sample peaks
}
\details{
The consensus set is the union of all peaks identified in any
  sample that overlap a peak in another sample.
}
