% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_peaks.R
\name{consensus_peaks_from_json}
\alias{consensus_peaks_from_json}
\title{Consensus peaks from a JSON file}
\usage{
consensus_peaks_from_json(json_file_path)
}
\arguments{
\item{json_file_path}{character, the path to a JSON file}
}
\value{
GRanges, the consensus peak set
}
\description{
As `consensus_peaks` but takes as input the path to a JSON file
  containing paths to peaks files instead of a list containing peaks
}
