% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_components_analysis.R
\name{coordinates_by_treatment}
\alias{coordinates_by_treatment}
\title{coordinates by treatment}
\usage{
coordinates_by_treatment(coord)
}
\arguments{
\item{coord}{two-column matrix of principal component coordinates}
}
\value{
list of two-column matrices, one per treatment
}
\description{
organize PCA coordinates by treatment
}
