% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrix.R
\name{count_dgelist}
\alias{count_dgelist}
\title{DGEList of read counts}
\usage{
count_dgelist(raw_counts, group)
}
\arguments{
\item{raw_counts}{matrix of raw counts}

\item{group}{group vector}
}
\value{
DGEList object representing read counts
}
\description{
generate a DGEList from raw read counts
}
