% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_counts.R
\name{edger_counts}
\alias{edger_counts}
\title{edgeR counts}
\usage{
edger_counts(peaks, reads_file_paths)
}
\arguments{
\item{peaks}{Granges object representing consensus peaks}

\item{reads_file_paths}{list or named character vector giving paths to BAM
files}
}
\value{
DGEList object for use with edgeR
}
\description{
read counts for input into edgeR
}
