% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Explore ATAC-seq data}
\usage{
explore(json_file_path, output_prefix, treatment_groups = list(),
  cores = 1)
}
\arguments{
\item{json_file_path}{path to a JSON file providing data details}

\item{output_prefix}{character, a prefix for output files}

\item{treatment_groups}{list providing groups of treatments to be compared}

\item{cores}{integer, max number of cores to use}
}
\description{
perform an exploratory analysis
}
