% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_components_analysis.R
\name{exploreatacseq_color_palette}
\alias{exploreatacseq_color_palette}
\title{tailored color palette}
\usage{
exploreatacseq_color_palette(order = "categorical")
}
\arguments{
\item{order}{set to either "categorical" or "sequential", determines
ordering of the palette}
}
\value{
character vector defining color palette
}
\description{
a color palette for PCA plots
}
