% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{extract_batch_vector}
\alias{extract_batch_vector}
\title{extract batch}
\usage{
extract_batch_vector(input_list)
}
\arguments{
\item{input_list}{list derived from input JSON}
}
\value{
character vector of batch
}
\description{
extract batch vector from input
}
