% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_counts.R
\name{extract_group_vector}
\alias{extract_group_vector}
\title{extract group vector}
\usage{
extract_group_vector(reads_file_paths)
}
\arguments{
\item{reads_file_paths}{list containing the reads file paths for each group}
}
\value{
a group vector for DGEList
}
\description{
extract a group vector (for DGEList) from the read file paths
  data
}
