% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{extract_peaks_paths_by_sample}
\alias{extract_peaks_paths_by_sample}
\title{extract peaks paths by sample}
\usage{
extract_peaks_paths_by_sample(input_list)
}
\arguments{
\item{input_list}{list derived from input JSON}
}
\value{
list of peaks file paths by sample
}
\description{
extract peaks paths from input
}
