% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{extract_reads_file_paths}
\alias{extract_reads_file_paths}
\title{extract reads file paths}
\usage{
extract_reads_file_paths(input_list)
}
\arguments{
\item{input_list}{list derived from input JSON}
}
\value{
character vector of reads file paths
}
\description{
extract reads file paths from input
}
