% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrix.R
\name{extract_sample_vector}
\alias{extract_sample_vector}
\title{extract sample vector}
\usage{
extract_sample_vector(count_matrix)
}
\arguments{
\item{count_matrix}{matrix of read counts}
}
\value{
character vector indicating sample
}
\description{
extract sample vector from count matrix
}
