% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{extract_tss_enrichments}
\alias{extract_tss_enrichments}
\title{extract TSS enrichments}
\usage{
extract_tss_enrichments(input_list)
}
\arguments{
\item{input_list}{list derived from input JSON}
}
\value{
numeric vector of TSS enrichments
}
\description{
extract TSS enrichment vector from input
}
