% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_peaks.R
\name{filter_peaks}
\alias{filter_peaks}
\title{Filter peaks}
\usage{
filter_peaks(peaks)
}
\arguments{
\item{peaks}{GRanges object representing peaks}
}
\value{
GRanges, the filtered peaks
}
\description{
filter out peaks that are greater than 3kb in length or are
    non-autosomal
}
