% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{generate_pca_plots}
\alias{generate_pca_plots}
\title{generate PCA plots}
\usage{
generate_pca_plots(counts, output_prefix, treatment,
  treatment_groups = list())
}
\arguments{
\item{counts}{matrix of read counts}

\item{output_prefix}{prefix for output files}

\item{treatment}{character vector indicating treatment}

\item{treatment_groups}{list of treatment groups}
}
\description{
plot PCA of the read counts
}
