% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{generate_umap_plots}
\alias{generate_umap_plots}
\title{generate UMAP plots}
\usage{
generate_umap_plots(counts, output_prefix, sample, treatment,
  treatment_groups = list())
}
\arguments{
\item{counts}{matrix of read counts}

\item{output_prefix}{prefix for output files}

\item{sample}{character vector indicating sample}

\item{treatment}{character vector indicating treatment}

\item{treatment_groups}{list of treatment groups}
}
\description{
plot UMAP of the read counts
}
