% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrix.R
\name{get_raw_counts}
\alias{get_raw_counts}
\title{Matrix of raw read counts}
\usage{
get_raw_counts(peaks, reads_file_paths, cores = 1)
}
\arguments{
\item{peaks}{Granges object representing consensus peaks}

\item{reads_file_paths}{list or named character vector giving paths to BAM
files}

\item{cores}{number of cores to use}
}
\value{
matrix of read counts
}
\description{
produce a matrix of raw read counts per peak
}
