% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_utils.R
\name{locate_SCATEData_files}
\alias{locate_SCATEData_files}
\title{locate SCATEData files}
\usage{
locate_SCATEData_files()
}
\value{
list of paths to SCATEData BAM files
}
\description{
locate BAM files from the SCATEData package
}
