% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_utils.R
\name{map_paths_to_samples}
\alias{map_paths_to_samples}
\title{map SCATEData file paths to sample types}
\usage{
map_paths_to_samples()
}
\value{
list mapping file paths to sample types, sorted by sample types
}
\description{
map SCATEData file paths to sample types
}
