% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_peaks.R
\name{median_peak_length}
\alias{median_peak_length}
\title{Median Peak Length}
\usage{
median_peak_length(peaks)
}
\arguments{
\item{peaks}{Granges object representing peaks}
}
\value{
integer, the median peak length
}
\description{
get the median peak length
}
