% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrix.R
\name{most_variable_peaks}
\alias{most_variable_peaks}
\title{most variable peaks}
\usage{
most_variable_peaks(count_matrix, n = 1e+05)
}
\arguments{
\item{count_matrix}{numeric matrix representing transformed counts}

\item{n}{max number of peaks to keep}
}
\value{
numeric matrix representing counts for most variable peaks
}
\description{
extract the most variable peaks from the count matrix
}
