% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{parse_input_data}
\alias{parse_input_data}
\title{parse input data}
\usage{
parse_input_data(input_list)
}
\arguments{
\item{input_list}{list derived from input JSON}
}
\value{
list of useful vectors
}
\description{
parse input data into useful vectors
}
