% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{parse_json}
\alias{parse_json}
\title{parse input JSON}
\usage{
parse_json(json_file_path)
}
\arguments{
\item{input_list}{path to input JSON file}
}
\value{
list of useful vectors
}
\description{
parse input JSON into useful vectors
}
