% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_peaks.R
\name{peaks_by_sample}
\alias{peaks_by_sample}
\title{Peaks by sample}
\usage{
peaks_by_sample(peaks_paths_by_sample)
}
\arguments{
\item{peaks_paths_by_sample}{a list of character vectors containing paths of
peaks files for each sample}
}
\value{
list of GRanges objects representing the peaks for each sample
}
\description{
Get a GRanges object containing a set of peaks for each input
  sample
}
