% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_peaks.R
\name{read_peaks}
\alias{read_peaks}
\title{Read peaks from a file}
\usage{
read_peaks(peaks_file_path)
}
\arguments{
\item{peaks_file_path}{character, path to the peaks file}
}
\value{
GRanges object representing the peaks
}
\description{
Read peaks from a BED or similar file into a GRanges object
}
