% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCATEData_accession_sample.R
\docType{data}
\name{SCATEData_accession_sample}
\alias{SCATEData_accession_sample}
\title{Table mapping SCATEData accession numbers to sample type}
\format{
A data frame with 20 rows and 2 variables:
\describe{
    \item{accession} SRA accession number
    \item{sample} Sample type
}
}
\source{
\href{https://www.ncbi.nlm.nih.gov/sra/}{NCBI Sequence Read Archive}
}
\usage{
data(SCATEData_accession_sample)
}
\description{
The bioconductor package SCATEData includes 20 BAM files containing
single-cell ATAC-seq data. Each dataset is named by its NCBI Sequence Read
Archive accession number, "SRR1779746.bam". This table lists the sample type
of each dataset.
}
\references{
Zhicheng Ji, Weiqiang Zhou, Wenpin Hou, Hongkai Ji,
\href{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-020-02075-3}{Single-cell ATAC-seq Signal Extraction and Enhancement with SCATE},
Genome Biol 21, 161 (2020)
}
\keyword{datasets}
