% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SCATEData_sample}
\alias{SCATEData_sample}
\title{Vector mapping SCATEData accession numbers to sample type}
\format{
A character vector with 20 values
}
\source{
\href{https://www.ncbi.nlm.nih.gov/sra/}{NCBI Sequence Read Archive}
}
\usage{
data(SCATEData_sample)
}
\description{
The bioconductor package SCATEData includes 20 BAM files containing
single-cell ATAC-seq data. Each dataset is named by its NCBI Sequence Read
Archive accession number, "SRR1779746.bam". This vector maps the accession
number to the sample type of each dataset.
}
\references{
Zhicheng Ji, Weiqiang Zhou, Wenpin Hou, Hongkai Ji,
\href{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-020-02075-3}{Single-cell ATAC-seq Signal Extraction and Enhancement with SCATE},
Genome Biol 21, 161 (2020)
}
\keyword{datasets}
