% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrix.R
\name{transform_counts}
\alias{transform_counts}
\title{apply transformations to read counts}
\usage{
transform_counts(count_dgelist)
}
\arguments{
\item{count_dgelist}{DGEList representing read counts}
}
\value{
numeric matrix representing transformed counts
}
\description{
apply voom and remove batch effects
}
