% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_components_analysis.R
\name{two_principal_components}
\alias{two_principal_components}
\title{two principal components}
\usage{
two_principal_components(count_matrix)
}
\arguments{
\item{count_matrix}{matrix of read counts}
}
\value{
matrix with two columns giving the principal component coordinates
}
\description{
compute two principal components for the count matrix
}
