% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_peaks.R
\name{unify_granges}
\alias{unify_granges}
\title{Unify GRanges}
\usage{
unify_granges(x, ignore_strand = FALSE)
}
\arguments{
\item{x}{list of GRanges objects}

\item{ignore_strand}{logical, if TRUE strand information will be ignored}
}
\value{
GRanges, the union of the GRanges in x
}
\description{
Union of several GRanges objects
}
