% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_peaks.R
\name{unify_peaks}
\alias{unify_peaks}
\title{Unify peaks}
\usage{
unify_peaks(peaks_file_paths)
}
\arguments{
\item{peaks_file_paths}{character, paths to the files containing the peaks}
}
\value{
GRanges, the union of peaks from the input files
}
\description{
Create a GRanges object representing the union of peaks from
 one or more input files
}
