% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrix.R
\name{count_dgelist}
\alias{count_dgelist}
\title{DGEList of read counts}
\usage{
count_dgelist(peaks, reads_file_paths, group, cores = 1)
}
\arguments{
\item{peaks}{Granges object representing consensus peaks}

\item{reads_file_paths}{list or named character vector giving paths to BAM
files}
}
\value{
DGEList object representing read counts
}
\description{
generate a DGEList from input peaks and BAM files
}
