% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Explore ATAC-seq data}
\usage{
explore(json_file_path, output_prefix, treatment_groups = list(),
  labels = FALSE, n_neighbors = 15, metric = "euclidean",
  n_pc = NULL, write_counts = FALSE, cores = 1, pca = TRUE,
  umap = FALSE, palette_order = "categorical")
}
\arguments{
\item{json_file_path}{path to a JSON file providing data details}

\item{output_prefix}{character, a prefix for output files}

\item{treatment_groups}{list providing groups of treatments to be compared}

\item{labels}{if TRUE, text labels will be added to points in all plots}

\item{n_neighbors}{size of local neighborhood for umap, see ?uwot::umap}

\item{metric}{distance metric for umap, see ?uwot::umap}

\item{n_pc}{number of principal components to pass to umap}

\item{write_counts}{logical, if TRUE the read count matrix will be written
to disk as a TSV file}

\item{cores}{integer, max number of cores to use}

\item{palette_order}{ordering of color palette, eithier "categorical" or
"sequential"}
}
\description{
perform an exploratory analysis
}
