% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Explore ATAC-seq data}
\usage{
explore(
  input_data,
  output_prefix,
  treatment_groups = list(),
  labels = FALSE,
  n_peaks = 1e+05,
  write_raw_counts = FALSE,
  write_transformed_counts = FALSE,
  write_counts = FALSE,
  cores = 1,
  palette_order = "categorical",
  palette = NULL
)
}
\arguments{
\item{input_data}{list contatining data details or character path to a JSON
file}

\item{output_prefix}{character, a prefix for output files}

\item{treatment_groups}{list providing groups of treatments to be compared}

\item{labels}{if TRUE, text labels will be added to points in all plots}

\item{n_peaks}{integer, number of the top most variable peaks to include in
the analysis}

\item{write_raw_counts}{logical, if TRUE the raw read count matrix will be
written to disk as a TSV file}

\item{write_transformed_counts}{logical, if TRUE the transformed read count
matrix will be written to disk as a TSV file}

\item{write_counts}{equivalent to `write_transformed_counts`, included to
preserve compatibility}

\item{cores}{integer, max number of cores to use}

\item{palette_order}{ordering of color palette, eithier "categorical" or
"sequential"}

\item{palette}{the color palette}
}
\description{
perform an exploratory analysis
}
\details{
The analysis consists of the following steps:
}
