% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Explore ATAC-seq data}
\usage{
explore(input_data, output_prefix, treatment_groups = list(),
  labels = FALSE, n_peaks = 1e+05, n_neighbors = 15,
  metric = "euclidean", n_pc = NULL, write_counts = FALSE,
  cores = 1, pca = TRUE, umap = FALSE,
  palette_order = "categorical", palette = NULL)
}
\arguments{
\item{input_data}{list contatining data details or character path to a JSON
file}

\item{output_prefix}{character, a prefix for output files}

\item{treatment_groups}{list providing groups of treatments to be compared}

\item{labels}{if TRUE, text labels will be added to points in all plots}

\item{n_peaks}{integer, number of the top most variable peaks to include in
the analysis}

\item{n_neighbors}{size of local neighborhood for umap, see ?uwot::umap}

\item{metric}{distance metric for umap, see ?uwot::umap}

\item{n_pc}{number of principal components to pass to umap}

\item{write_counts}{logical, if TRUE the read count matrix will be written
to disk as a TSV file}

\item{cores}{integer, max number of cores to use}

\item{pca}{logical, if true perform PCA analysis}

\item{umap}{logical, if true perform UMAP analysis}

\item{palette_order}{ordering of color palette, eithier "categorical" or
"sequential"}

\item{palette}{the color palette}
}
\description{
perform an exploratory analysis
}
