% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{extract_treatment_vector}
\alias{extract_treatment_vector}
\title{extract treatment vector}
\usage{
extract_treatment_vector(counts)
}
\arguments{
\item{counts}{matrix of read counts}
}
\value{
character vector indicating treatment
}
\description{
extract treatment vector from count matrix
}
