% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{generate_pca_plots}
\alias{generate_pca_plots}
\title{generate PCA plots}
\usage{
generate_pca_plots(counts, output_prefix, treatment = NULL,
  treatment_groups = list(), labels = FALSE,
  palette_order = "categorical")
}
\arguments{
\item{counts}{matrix of read counts}

\item{output_prefix}{prefix for output files}

\item{treatment}{character vector indicating treatment, if NULL, it will be
inferred from the read count matrix}

\item{treatment_groups}{list of treatment groups}

\item{labels}{if TRUE, text labels will be added to points in all plots}

\item{palette_order}{ordering of color palette, eithier "categorical" or
"sequential"}
}
\description{
plot PCA of the read counts
}
