% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{generate_umap_plots}
\alias{generate_umap_plots}
\title{generate UMAP plots}
\usage{
generate_umap_plots(counts, output_prefix, sample = NULL,
  treatment = NULL, treatment_groups = list(), labels = FALSE,
  n_neighbors = 15, metric = "euclidean", n_pc = NULL, cores = 1)
}
\arguments{
\item{counts}{matrix of read counts}

\item{output_prefix}{prefix for output files}

\item{sample}{character vector indicating sample, if NULL, it will be
inferred from the read count matrix}

\item{treatment}{character vector indicating treatment, if NULL, it will be
inferred from the read count matrix}

\item{treatment_groups}{list of treatment groups}

\item{labels}{if TRUE, text labels will be added to points in all plots}

\item{n_neighbors}{size of local neighborhood for umap, see ?uwot::umap}

\item{metric}{distance metric for umap, see ?uwot::umap}

\item{n_pc}{pass the first n principal components to UMAP}

\item{cores}{number of cores to use}
}
\description{
plot UMAP of the read counts
}
