% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_components_analysis.R
\name{plot_pca}
\alias{plot_pca}
\title{plot principal components}
\usage{
plot_pca(pca, draw_lines = list(), labels = FALSE,
  palette = brewer.pal(9, "Set1")[c(2, 1, 3:5, 7:9, 6)])
}
\arguments{
\item{pca}{list with class "prcomp"}

\item{draw_lines}{list of treatment groups to draw lines through}

\item{labels}{if TRUE, text labels will be added to the points}
}
\description{
generate a PCA plot
}
