% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{preprocess}
\alias{preprocess}
\title{preprocess ATAC-seq data}
\usage{
preprocess(input_data, n_peaks = 1e+05, cores = 1)
}
\arguments{
\item{input_data}{list contatining data details or character path to a JSON
file}

\item{n_peaks}{integer, number of the top most variable peaks to include in
the analysis}

\item{cores}{integer, max number of cores to use}
}
\value{
list giving the median peak length and the matrix of read counts
}
\description{
preprocess multiple ATAC-seq datasets into a read count matrix
}
