% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrix.R
\name{transform_counts}
\alias{transform_counts}
\title{apply transformations to read counts}
\usage{
transform_counts(count_dgelist, batch = NULL, covariates = NULL)
}
\arguments{
\item{count_dgelist}{DGEList representing read counts}

\item{batch}{factor or vector indicating batches}

\item{covariates}{matrix or vector of numeric covariates}
}
\value{
numeric matrix representing transformed counts
}
\description{
apply voom to input read counts (DGEList) and remove batch
    effects
}
